requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==5.0.9
enum34==1.1.10
idna==2.10
iniconfig==1.1.1
lxml==4.6.3
namedlist==1.8
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
selenium==3.141.0
self==2020.12.3
six==1.16.0
toml==0.10.2
urllib3==1.26.5
